Breaking changes
The first of the breaking changes is the addition of uppercase letters in
`LETTERS`. The second breaking change is fixing the overflow caused by
using `math.MaxInt64` as the argument for `rand.Intn`. Changed to using
`rand.Int63n`. Re-added `crypto/md5` and `crypto/sha1` hash functions as
options for `-a`. Updated examples and all READMEs.